Create README.md
Background: 
An institutional investor approached us to help him determine the average number of days it takes for the CBOE Volatility Index (VIX) to reach a certain number of points from a starting point, as well as the probability of the index reaching that number from that starting point. The investor had over 30 years of VIX data, spanning more than 8000 rows.

Problem: 
While the dataset itself was fairly standard, the task required complex data manipulation to achieve the desired results. We initially looked to Python for a solution, but the amount of coding required would have been significant.

Solution: 
After much brainstorming, we discovered an alternative solution through an Excel tool called Power Query. With this tool, we were able to extract the necessary data and calculate the probabilities and average number of days required for the VIX to reach a specific level from a given starting point.
The client was thrilled with the results, as this information greatly assisted him in making informed trading decisions. Additionally, we gained valuable insights into the capabilities of Power Query and its potential for use in future projects.